refactor: separate task execution from collection tracking
