+++ src/api/index.ts
@@ -1,0 +1,11 @@
+export { task, type TaskOptions } from "./task";
+export {
+  all,
+  forEach,
+  type EffectExecutionOptions,
+  type EffectAllExecutionOptions,
+  type AllOptions,
+  type AllReturn,
+  type ForEachExecutionOptions,
+  type ForEachOptions,
+} from "./collections";
