+++ src/api/infer-total.ts
